mi-codigo-rag/requirements.txt
functions-framework==3.*
google-cloud-aiplatform==1.56.0
google-cloud-documentai
google-cloud-storage
tenacity
langchain-google-vertexai==1.0.3
langchain==0.1.20
langchain-community==0.0.38
google-cloud-logging